+++ classes/BIM/App/Clubs.php
@@ -29,11 +29,23 @@ class BIM_App_Clubs extends BIM_App_Base{
         if( $club->isOwner($ownerId) ){
             $invited = $club->invite( $users, $nonUsers );
             if( $invited ){
+                self::postInvitationEvents($clubId, $ownerId, $users, $nonUsers);
                 self::notifyInvitees($clubId, $users, $nonUsers);
                 //BIM_Jobs_Clubs::queueNotifyInvitees($clubId, $users, $nonUsers);
             }
         }
         return $invited;
+    }
+
+    public static function postInvitationEvents( $clubId, $actorMemberId, $invitees, $nonUsers ) {
+        if ( count($invitees) >= 1 && $clubId && $actorMemberId ) {
+            $eventDispatcher = new BIM_EventDispatcher_Club();
+            if ( is_object($eventDispatcher) ) {
+                foreach ( $invitees as $inviteeMemberId ) {
+                    $eventDispatcher->invitationToMember($clubId, $actorMemberId, $inviteeMemberId);
+                }
+            }
+        }
     }
 
     public static function notifyInvitees( $clubId, $users, $nonUsers ) {
@@ -105,8 +117,20 @@ class BIM_App_Clubs extends BIM_App_Base{
         $joined = false;
         if( $club->isExtant() ){
             $joined = $club->join( $userId );
+            if( $joined ){
+                self::postJoinEvent($clubId, $userId);
+            }
         }
         return $joined;
+    }
+
+    public static function postJoinEvent( $clubId, $joinerId) {
+        if ( $clubId && $joinerId) {
+            $eventDispatcher = new BIM_EventDispatcher_Club();
+            if ( is_object($eventDispatcher) ) {
+                $eventDispatcher->memberJoined($clubId, $joinerId);
+            }
+        }
     }
 
     public static function quit( $clubId, $userId ){