+++ classes/BIM/Event/Club.php
@@ -1,6 +1,0 @@
-<?php
-class BIM_Event_Club extends BIM_Event_Base {
-
-}
-
-?>